frontend/requirements.txt
streamlit==1.24.0
Pillow==9.5.0
websockets==10.4
requests==2.31.0